chainparams: Change nChainTx to uint64_t
Summary:
Also update types of assumeutxo chainparams and some related local variables for
consistency.


This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/29656 | core#29656]]
https://github.com/bitcoin/bitcoin/pull/29656/changes/dc2938e9799d79696d1db2438ef33d90542d984b
https://github.com/bitcoin/bitcoin/pull/29656/changes/72e5d1be1f4491565249d43e836ee42cfd858866

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20656

+++ src/blockindex.h
@@ -64,8 +64,7 @@ public:
     //! This value will be non-zero if this block and all previous blocks back
     //! to the genesis block or an assumeutxo snapshot block have reached the
     //! VALID_TRANSACTIONS level.
-    //! Change to 64-bit type when necessary; won't happen before 2030
-    unsigned int nChainTx{0};
+    uint64_t nChainTx{0};
 
     //! Verification status of this block. See enum BlockStatus
     BlockStatus nStatus GUARDED_BY(::cs_main){};
@@ -135,7 +134,7 @@ public:
     /**
      * Get the number of transaction in the chain so far.
      */
-    int64_t GetChainTxCount() const { return nChainTx; }
+    uint64_t GetChainTxCount() const { return nChainTx; }
 
     /**
      * Check whether this block and all previous blocks back to the genesis
